refactor(domparser): split index.ts into a barrel
Pure move: the node classes go to src/dom/, the XML scanner to
src/xml/parse-xml.ts, DOMParser to src/dom-parser.ts, and src/index.ts
becomes re-exports only (root AGENTS.md — monolithic entry points).

No behaviour change: src/index.spec.ts is untouched and its 28 XML cases
stay green on node and gjs. The node-type constants move with the classes
so later call sites read as the spec does instead of comparing integers.
